% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getResponsePredictors.R
\name{getLHS.RHS}
\alias{getLHS.RHS}
\alias{lhs.form}
\alias{rhs.form}
\title{These functions extract the response or predictors of a formular}
\usage{
lhs.form(formula)

rhs.form(formula)
}
\arguments{
\item{formula}{R formula}
}
\value{
response or predictor variables
}
\description{
These functions extract the response or predictors of a formular
}
\examples{
form <- as.formula("y ~ V1 + V2")
resp <- lhs.form(form)
rhs.vars <- rhs.form(form)
}

