% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kernels.R
\name{kernels}
\alias{Fourier}
\alias{HardLimit}
\alias{HypTan}
\alias{Kernel.ELR}
\alias{MultiQuad}
\alias{Weights.ELR}
\alias{kernels}
\alias{randomMatrix.ELR}
\alias{rbf}
\alias{sigmoid.kernel.ELR}
\alias{wavelet.hyperbolic.sine.kernel}
\title{Functions to compute randomized feature space kernels.}
\usage{
sigmoid.kernel.ELR(dat, p, W)

wavelet.hyperbolic.sine.kernel(dat, p, W, ratio = 0.3)

rbf(dat, p, W)

HypTan(dat, p, W)

Fourier(dat, p, W)

HardLimit(dat, p, W)

MultiQuad(dat, p, W)

Kernel.ELR(dat, p, W, ken = "sigmoid")

Weights.ELR(d, p, ken, seed = NULL)

randomMatrix.ELR(nCols, nRows, minValue, maxValue)
}
\arguments{
\item{dat}{A matrix/data frame}

\item{W}{Random Weight matrix as computed by the function \code{\link{Weights.ELR}}}

\item{ratio}{mixture proportion for hybrid kernel}

\item{ken}{kernel function}

\item{d,p}{Dimension of data/random feature space}

\item{seed}{The value of .Random.seed for which ELM was trained. Set to NULL during training}

\item{nCols,nRows}{number of columns and rows of random weight matrix}

\item{minValue,maxValue}{minimum and maximum value for uniform distribution}
}
\value{
random feature space kernel matrix or random weight matrix
}
\description{
Though these functions are exported, the user will not necessary use them directly.
They are used internally by other functions and packages
}

