% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.ELR.R
\name{predict.ELR}
\alias{predict.ELM}
\alias{predict.ELR}
\title{predict.ELR}
\usage{
\method{predict}{ELR}(object, newdata, type = c("prob", "class")[1], ...)

\method{predict}{ELM}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted model from \code{\link{ELR}}.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for
class labels.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with items
\item{prob}{predicted class probabilities}
\item{class}{predicted class memberships obtained by thresholding
class probabilities at the prevalence rate of the positive class}
}
\description{
Make predictions using a fitted ELR model




}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

