% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Train.ELR.R
\name{Train.ELR}
\alias{Train.ELR}
\alias{TrainELM}
\alias{TrainELR}
\alias{TrainIRLSELR}
\title{ELR}
\usage{
TrainELR(X, Y, p, gamma, ken, ...)

TrainIRLSELR(X, Y, p, gamma, ken, max.iter, tol)

TrainELM(X, Y, p, gamma, ken)
}
\arguments{
\item{X}{Matrix of predictors}

\item{Y}{0,1 binary response matrix}

\item{p}{Dimension of random feature space}

\item{gamma}{Numeric regularization parameter}

\item{ken}{Type of kenel function}

\item{...}{Further arguments passed to or from other methods.}

\item{max.iter}{Maximum number of iteration}

\item{tol}{Convergence tolerance}
}
\value{
A list with items
\item{beta}{estimated parameters}
\item{p}{Dimension of random feature space}
\item{W}{Random weight matrix}
\item{ken}{kernel function}
}
\description{
Train least squares ELR (LS-ELR) and iterative re-weighted least squares ELR
 (IRLS-ELR).
}

