% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{utility.functions}
\alias{hprod}
\alias{invHypSine}
\alias{sigminus}
\alias{sigmoid}
\alias{sigplus}
\alias{tr}
\alias{utility.functions}
\alias{wavelet}
\title{Utility functions.}
\usage{
sigmoid(x)

sigminus(x)

sigplus(x)

tr(x)

hprod(x, y)

wavelet(x)

invHypSine(x)
}
\arguments{
\item{x,y}{A numeric or character vector depending on the function}
}
\value{
A numeric or character vector depending on the function
}
\description{
These functions are not meant to be directly called by the user. They
are used internally by other functions and packages that may depend on
ELR: sigmoid is the sigmoid function,
tr computes the trace of a matrix, hprod computes the hadamard product
}

